Address review nits: move test into test_fsdp_hybrid_shard.py
- Move test from standalone test_hsdp_sync_module_states.py into
  test_fsdp_hybrid_shard.py
- Define ModelWithBuffer as inner class of TestHSDPSyncModuleStates
